==> Part 1) a) Brand review.ipynb <==
# Part 1 - Determine which brands require further analysis

### Import the required packages

import pandas as pd
from bs4 import BeautifulSoup
import requests
import time

### Import list of car manufacturers listed on TradeMe:

This list was created from copying the inner html element of the dropdown menu on the TradeMe website, removing the surrounding code, and saving to a .csv file.
Only small changes were needed to formatting, e.g. 'Aston Martin' became 'Aston-Martin'.

car_makes_df = pd.read_csv('csv_files/CarManufacturers.csv')

### Check list of manufacturers:

car_makes_df.head()

### Create list of URL addresses to retrieve listing data from, using the list of manufacturers above.

# Head URL common to all:
url = f'https://www.trademe.co.nz/a/motors/cars/'

# Empty list to hold final result
url_list = []

# Loop to concatenate head URL with manufactuerer name and add result to the url_list
for make in car_makes_df['Make']:
    url_list.append(url+make)

# Check result on first three items in list:
print(url_list[0:3])

### Create empty dictionary to hold results from webscraper

results = {}

### Create loop to search each URL from the list, retrieve the number of listings based on the `<div>` class, and save the result to the results dictionary

for url in url_list:
    # Create variable for brand name based on URL ending, Set source of data
    brand = url[40:] 
    source = requests.get(url).text
    # Parse source data
    soup = BeautifulSoup(source, 'lxml')
    # Search source for div class identified:
    match = soup.find('div', class_="tm-motors-search-results__sort-and-view-options")
    h3 = soup.find('h3')
    # Print results to terminal and save result to the dictionary created above 
    try:
        print(brand + " : " + h3.text)
        results.update({brand: h3.text})
    except:
        print(brand + ': Error')
        results.update({brand: 'Error'})
    # Pause, to avoid spamming TradeMe server
    time.sleep(3.5)

### Create a dataframe with the listing count, from the dictionary created by the loop above.

count_listings_df = pd.DataFrame.from_dict(results, orient='index')

count_listings_df.columns = ['Results']

### Check result:

count_listings_df.head()

### Save the result to a .csv file for safekeeping:

count_listings_df.to_csv('csv_files/brand_results.csv')
