==> Part 1) b) Brand listing results.ipynb <==
# Part 1) b) Review listing count by manufacturer

Here we are going to sort the dataframe of the results in Part 1, to see the spread of manufacturer listings.

import pandas as pd
import numpy as np

### Import .csv file

df = pd.read_csv('csv_files/brand_results.csv')

### Check format

df.head()

### Rename 'unnamed' column and set index 

df.rename(columns = {'Unnamed: 0':'Brand'}, inplace = True)

df.set_index('Brand', inplace=True)

### Create new column for listing count as a numerical field

df['Count'] = df['Results'].str[9:-9]

df['Count'] = df['Count'].str.replace(',','')

df['Count'] = df['Count'].replace('', np.nan, regex=True)

df['Count'] = df['Count'].astype(float)

### Sort dataframe column by listing count

df.sort_values(by=['Count'], ascending=False).head()

### Count total number of listings, and calculate percentage of total

df['Count'].sum()

df['Percentage'] = df['Count'] / 78549 * 100

df.sort_values(by=['Count'], ascending=False).head()

# Conclusion:

To produce statistcially signidficant results for this project, we need lots of homogenous data points (i.e. make/brand combinations) that repeat across the country.

Whilst it is possible for a manufacturer to produce a very high volumne of a small number of models, as the top 5 manufacturers make up nearly 50% of listings (48.5%), we can safely exclude the remaining brands entirely from future analysis. 

The most sold models would be expected to be included in this set, and should be representative of any national trends.
